refactored index into component names
