+++ server/client/src/components/channel/message/Message.js
@@ -1,7 +1,7 @@
 import React, { Component } from 'react';
 import {connect} from 'react-redux';
 
-import './index.css';
+import './message.css';
 
 class Message extends Component {
   render(){
